requirements.txt
sympy
numpy
segno
qrtools
opencv-python
opencv-python-headless
pyzbar
segno-pil